[svn r33] Fixed a bug with not allowing decimal points in the odometer reading. Don't show/hide the comment field--just let it be present but empty. Fixed the history layout so that the vehicle name (if present) is centered.

+++ trunk/src/com/evancharlton/mileage/FillUpView.java
@@ -105,7 +105,7 @@ public class FillUpView extends Activity {
 				}
 
 				try {
-					int mileage = Integer.parseInt(m_mileageEdit.getText().toString());
+					double mileage = Double.parseDouble(m_mileageEdit.getText().toString());
 					values.put(FillUps.MILEAGE, mileage);
 				} catch (NumberFormatException nfe) {
 					error = true;
